Render distance can now be changed at runtime before the programme starts instead of being determined at compile time but this also meant that I had to move the array of pointers to loaded chunks outside of the world struct

+++ game/header/terrainGen.h
@@ -2,4 +2,4 @@
 #include "./chunk.h"
 
 int generateChunk(fullChunk* chunk, world* loadedWorld);
-void initTerrainGen(world* loadedWorld, int seed);
+void initTerrainGen(world* loadedWorld, int64_t seed);
